// lib/api/generated/models/UpdateGroupViewModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateGroupViewModel = {
    name?: string | null;
    courseId?: string | null;
    topicId?: string | null;
    maxMembers?: number | null;
    startDate?: string | null;
    endDate?: string | null;
};